refactor: ViewController - CrewInfoView 분리
